requirements.txt
Flask==2.2.2
google-cloud-spanner==3.16.0
google-cloud-logging==3.2.2
gunicorn==20.1.0
Werkzeug==2.3.7
